Fix: 엘리베이터 알고리즘 수정
-엘리베이터 알고리즘 수정 (8층 초기화 시 특정 좌표에서 스폰, 아니면 엘리베이터로 이동)

+++ Source/Endless_Hotel/GameSystem/Anomaly/Anomaly_Generator.cpp
@@ -15,7 +15,7 @@ void AAnomaly_Generator::BeginPlay()
 
 	auto* Sub = GetGameInstance()->GetSubsystem<UAnomalyProgressSubSystem>();
 
-	uint8 IsNormal = FMath::RandRange(1, 10);
+	int32 IsNormal = FMath::RandRange(1, 10);
 	if (IsNormal > 8 || Sub->Floor == 9)
 	{
 		SpawnNormal();
